Move BraidTestProvider to separate entry (#1014)
