+++ lib/components/Accordion/AccordionItem.test.tsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom/extend-expect';
 import React, { useState } from 'react';
 import { render } from '@testing-library/react';
-import { BraidTestProvider, AccordionItem } from '..';
+import { BraidTestProvider } from '../../../test';
+import { AccordionItem } from '..';
 import { htmlToText } from '../../utils/htmlToText';
 
 describe('AccordionItem', () => {
