+++ lib/components/Checkbox/Checkbox.test.tsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom/extend-expect';
 import React, { ComponentProps, useState } from 'react';
 import { render } from '@testing-library/react';
-import { BraidTestProvider, Checkbox } from '..';
+import { BraidTestProvider } from '../../../test';
+import { Checkbox } from '..';
 import userEvent from '@testing-library/user-event';
 
 describe('Checkbox', () => {
