+++ lib/components/Link/Link.test.tsx
@@ -2,7 +2,8 @@ import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import { render } from '@testing-library/react';
 import { htmlToText } from '../../utils/htmlToText';
-import { BraidTestProvider, makeLinkComponent, LinkComponent, Link } from '..';
+import { BraidTestProvider } from '../../../test';
+import { makeLinkComponent, LinkComponent, Link } from '..';
 
 describe('Link', () => {
   it('should render a native link by default', () => {
