+++ lib/components/RadioGroup/RadioGroup.test.tsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom/extend-expect';
 import React, { useState } from 'react';
 import { render } from '@testing-library/react';
-import { BraidTestProvider, RadioGroup, RadioItem } from '..';
+import { BraidTestProvider } from '../../../test';
+import { RadioGroup, RadioItem } from '..';
 import userEvent from '@testing-library/user-event';
 
 const TestCase = () => {
