+++ lib/components/TextLinkButton/TextLinkButton.test.tsx
@@ -2,7 +2,8 @@ import '@testing-library/jest-dom/extend-expect';
 import React, { MouseEvent } from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { BraidTestProvider, TextLinkButton, Text, Actions } from '..';
+import { BraidTestProvider } from '../../../test';
+import { TextLinkButton, Text, Actions } from '..';
 
 const createMockClickHander = () =>
   jest.fn(
