+++ lib/components/Textarea/Textarea.test.tsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import { render } from '@testing-library/react';
-import { BraidTestProvider, Textarea } from '..';
+import { BraidTestProvider } from '../../../test';
+import { Textarea } from '..';
 
 describe('Textarea', () => {
   it('associates field with label correctly', () => {
