+++ lib/components/useBreakpoint/useBreakpoint.test.tsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import { renderHook } from '@testing-library/react-hooks';
-import { BraidTestProvider, useBreakpoint } from '..';
+import { BraidTestProvider } from '../../../test';
+import { useBreakpoint } from '..';
 
 describe('useResponsiveValue', () => {
   const testData = [
